SAT solvers - remove warnings V

+++ Hydra/external/satSolvers/MiniSat_d4/Solver.cpp
@@ -730,8 +730,7 @@ namespace minisat_d4 {
     };
     void Solver::reduceDB() {
         int i, j;
-        double extra_lim =
-                cla_inc / learnts.size();   // Remove any clause below this activity
+        float extra_lim = (float)(cla_inc / learnts.size());   // Remove any clause below this activity
 
         sort(learnts, reduceDB_lt(ca));
         // Don't delete binary or locked clauses. From the rest, delete clauses from
